requirements.txt
amqp==5.0.6
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Cython==0.29.24
idna==3.2
kombu==5.1.0
lxml==4.6.3
Pillow==8.3.1
prompt-toolkit==3.0.19
pytesseract==0.3.8
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tesseract==0.1.3
tqdm==4.61.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
